+++ device-management-application/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [[_TOC_]]
 
 ## [Unreleased]
+- [SW-19030] Add new commands DM_CMD_SET_VMIN_LUT and DM_CMD_GET_VMIN_LUT
 ### Added
 ### Changed
 ### Deprecated